Build on msvc x86-64.

+++ components/physics/bullet_driver/tests/shared.h
@@ -45,14 +45,14 @@ void dump_body_position (IRigidBody* body)
 
   math::vec3f world_position = body_transform.position;
 
-  for (size_t i = 0; i < 3; i++)
+  for (unsigned int i = 0; i < 3; i++)
     convert_float (world_position [i], 1);
 
   printf ("  position = %.0f; %.0f; %.0f\n", world_position.x, world_position.y, world_position.z);
 
   math::quatf world_orientation = body_transform.orientation;
 
-  for (size_t i = 0; i < 4; i++)
+  for (unsigned int i = 0; i < 4; i++)
     convert_float (world_orientation [i], 10);
 
   printf ("  orientation = %.1f; %.1f; %.1f; %.1f\n", world_orientation.x, world_orientation.y, world_orientation.z, world_orientation.w);
